Rename XOR* mnemonics to EOR*
Summary:
TCA mnemonics all start with an X, XOR was the only exception.  This
renaming makes it so that only TCA mnemonics start by an X.

On KV3-1 and KV3-2, we install synthetics so that XOR still works.
However, the KV4-1 will not expose a XOR mnemonic.

Test Plan: CI

Reviewers: csix, dsampaio, O81 llvm-coolidge!

Differential Revision: https://phab.kalray.eu/D33442