[kvx][NFC] Change `kvx-` prefix to `kvx-cos-`
Summary:
When setting a prefix to clang that is used to extract the triple
arch and os. Ommiting `cos` makes the driver to fail initializing
the ClusterOS and setting the correct include paths as well the
correct toolchain path, calling the system's assembler/linker
instead.

Test Plan: N/A

Reviewers: csix, O81 llvm-coolidge

Reviewed By: csix, O81 llvm-coolidge

Subscribers: echaloin

Maniphest Tasks: T21563

Differential Revision: https://phab.kalray.eu/D33609