[kvx][Driver] Always sets `-march` to assembler
Summary:
KVX can't really multilib between architectures.
This set's the `-march` argument passed to
the assembler as the same as passed to clang
driver, or "kv3-1" if omitted and ignores any
flags set for the libraries.

PS: I can't add a test as kvx-cos-as is not in most
dev's PATH and would fail.

Jira: https://kalray.atlassian.net/browse/TOOLCHAIN-145

Test Plan: N/A

Reviewers: csix, O81 llvm-coolidge

Reviewed By: csix, O81 llvm-coolidge

Differential Revision: https://phab.kalray.eu/D33643