[kvx] You may TailCall
Summary:
Implements TLI hooks that do check if a given function call can be
lowered to a tail-call depending if the middle-end has marked it as
tail-call and it's position in the functions and uses.
All the logic of tail-lowering was existed but it wouldn't match in most
cases.

Jira: https://kalray.atlassian.net/browse/TOOLCHAIN-218

Test Plan: added/changed tests

Reviewers: csix, O81 llvm-coolidge!

Differential Revision: https://phab.kalray.eu/D34631