[kvx] FIX 4 x 32 bits register insert vector
Summary:
The patterns used to perform insert a 32 bit value to a register defined index
into a 4 elements vector were wrong. First the bit range of the upper element
was `31 - 61` instead of `32 - 61`. As well the selection condition of changing
or not the elements of the upper/lower half were inverted.

Jira: https://kalray.atlassian.net/browse/TOOLCHAIN-302

Test Plan: Changed test files

Reviewers: csix, O81 llvm-coolidge!

Differential Revision: https://phab.kalray.eu/D35394