[kvx] Don't interleave if not vectorized
Summary:
Prevents VPLan to interleave loops if it was not capable to vectorize loops.
It does, in many cases, duplicate loops with something not much more
efficient and exposes a instcombine bug. Once the IC bug is fixed we should
analyse the performance impact of it.

Jira: https://kalray.atlassian.net/browse/TOOLCHAIN-454

Test Plan: Changed files, llvm-test-suite

Reviewers: csix, O81 llvm-coolidge!

Differential Revision: https://phab.kalray.eu/D36337