[kvx][FIX UPSTREAM] Do no double-treat invokes
In 9fe78db4cd1c18fc36a2f2bcc849adf2941027ba adds nounwind inference for landingpads
that are not yet threated. However it does not differentiate which invokes have are
threated or not. This makes the same invoke being threated twice.
For the input code:
```
static int c;

struct A {
  A() { ++c; }
  A(const A&) { ++c; }
  ~A() { --c; }
};
int main() {
  try {
    throw new A();
  } catch (A *a) {delete a;}
  if (!c) printf("Base pointer ok!\n");
return 0;
```
It will overwrite in memory the landing pad of threating (A *) with the a pointer to -1.

```
  %call_site15 = getelementptr { ptr, i64, [4 x i64], ptr, ptr, [5 x ptr] }, ptr %fn_context, i32 0, i32 1, !dbg !38
  store volatile i64 2, ptr %call_site15, align 8, !dbg !38
  call void @llvm.eh.sjlj.callsite(i32 2)
  %call_site16 = getelementptr { ptr, i64, [4 x i64], ptr, ptr, [5 x ptr] }, ptr %fn_context, i32 0, i32 1, !dbg !38
  store volatile i64 -1, ptr %call_site16, align 8, !dbg !38
  invoke void @__cxa_throw(ptr nonnull %exception.reload, ptr nonnull @_ZTIP1A, ptr null) #12
          to label %unreachable unwind label %lpad3, !dbg !38
```
Here, the store to `call_site16` replaces the capturing of `A*` stored in `call_site15`.