[kvx] Implement scall builtin
Summary:
In the frontend, generate notail call by casting the scall 1st argument to a
pointer of addrspace 259

In the backend, at call lowering, look at the callee type: if it is a pointer
of addrspace 259, assume it is a SCALL.

We create a new ISD node named SCALL to differentiate calls and tailcalls from
system calls.

Add a new file KVXCommon.h to include definitions that are common to the
frontend and the backend. For now, just move the ADDRSPACE enum in there.

Also includes minor refactoring:
- use namespaces in CGBuiltin.cpp to avoid ambiguities between clang::KVX and llvm::KVX namespaces
- factorize list of physical registers defined in CALL/ICALL/SCALL

Test Plan: added frontend and backend CodeGen tests

Reviewers: dsampaio, O81 llvm-coolidge

Reviewed By: dsampaio, O81 llvm-coolidge

Subscribers: piannetta

Differential Revision: https://phab.kalray.eu/D36822