[kvx] Fix wrong type in ATOMIC_CMP_SWAP_WITH_SUCCESS legalization
The initial code was failing an assert in
SelectionDAGLegalize::LegalizeOp that checks whether the types remain
the same. The 2nd value of the ATOMIC_CMP_SWAP_WITH_SUCCESS node had a
type of i32 whereas we legalized it into a type of i64.

Fix the problem in buildCV2CmpSwap by using N->getValueType(1) instead
of N->getValueType(0).

Ref TOOLCHAIN-714

Link: https://gerrit.kalray.eu/id/I60cb8c3c62cb19542049923565810dca67e6bb59