[kvx][DOWNSTREAM] Python version hack
Ubuntu 18 latest official python is 3.6.9.
Upgrading it to 3.8 breaks the OS.
Installing 3.8 it and adding it to the path solves
the problem of compiling llvm. However cmake
won't detect 3.8 and will think the latest version
is insufficient. This hack relies in the fact that
python 3.8 will kick in as it is in the path at
tests execution time.

Implements: https://kalray.atlassian.net/browse/TOOLCHAIN-322
Link: https://gerrit.kalray.eu/id/I61af3038276ba821f20b011e940aa09aa25b9f11