[kvx] Yes, you may truncate tailcall
Allow taicalling when the result is an integer value that should be
truncated, such as callee returns i32 and caller returns the result
of callee truncated to i16. This is possible as the KVX calling
convention does not define values of the upper bits in arguments or
return values. It is the responsibility of the function actually
processing the data to actually extend the value in register.
Allowing i128 is also possible, as the resulting register value will sit
in the first of the 2 used register, turning into a `nop`.

Implements: https://kalray.atlassian.net/browse/TOOLCHAIN-797
Link: https://gerrit.kalray.eu/id/I27d4ca522bc72e00fbbea63bd0e21637344b84a5